+++ test/server/spec/bin/builder-core.spec.js
@@ -10,6 +10,7 @@
  * Tests _do_ however use real process `exec`'s. This _slightly_ slows things
  * down, but allows us to have some very real use case scenarios.
  */
+var fs = require("fs");
 var path = require("path");
 var chalk = require("chalk");
 
@@ -21,24 +22,22 @@ var run = require("../../../../bin/builder-core");
 
 var base = require("../base.spec");
 
-// Helpers
-// **Note**: It would be great to just stub stderr, stdout in beforeEach,
-// but then we don't get test output. So, we manually stub with this wrapper.
-var stdioWrap = function (fn) {
-  return function (done) {
-    base.sandbox.stub(process.stdout, "write");
-
-    var _done = function (err) {
-      process.stdout.write.restore();
-      done(err);
-    };
-
+// Read file, do assert callbacks, and trap everything, calling `done` at the
+// end. A little limited in use as it's the *last* thing you can call in a
+// given test, but we can abstract more later if needed.
+var readFile = function (filename, callback, done) {
+  base.mockFs.restore();
+  fs.readFile(filename, { encoding: "utf8" }, function (err, data) {
     try {
-      return fn(_done);
-    } catch (err) {
-      return _done(err);
+      if (!err) {
+        callback(data); // eslint-disable-line callback-return
+      }
+    } catch (assertErr) {
+      err = assertErr;
     }
-  };
+
+    done(err);
+  });
 };
 
 describe("bin/builder-core", function () {
@@ -121,7 +120,8 @@ describe("bin/builder-core", function () {
 
   describe("builder --version", function () {
 
-    it("runs version", stdioWrap(function (done) {
+    it("runs version", function (done) {
+      base.sandbox.spy(process.stdout, "write");
       base.sandbox.spy(Task.prototype, "version");
       base.mockFs({
         "package.json": "{}"
@@ -137,7 +137,7 @@ describe("bin/builder-core", function () {
 
         done();
       });
-    }));
+    });
 
   });
 
@@ -252,12 +252,12 @@ describe("bin/builder-core", function () {
 
   describe("builder run", function () {
 
-    it("runs a <root>/package.json command", stdioWrap(function (done) {
+    it("runs a <root>/package.json command", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         "package.json": JSON.stringify({
           "scripts": {
-            "bar": "echo BAR_TASK"
+            "bar": "echo BAR_TASK >> stdout.log"
           }
         }, null, 2)
       });
@@ -268,42 +268,20 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.run).to.be.calledOnce;
-        expect(process.stdout.write).to.be.calledWithMatch("BAR_TASK");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data).to.contain("BAR_TASK");
+        }, done);
       });
 
-    }));
+    });
 
-    it("runs a with an unlimited buffer", stdioWrap(function (done) {
+    it("runs with quiet log output", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         "package.json": JSON.stringify({
           "scripts": {
-            "bar": "echo BAR_TASK"
-          }
-        }, null, 2)
-      });
-
-      run({
-        argv: ["node", "builder", "run", "--unlimited-buffer", "bar"]
-      }, function (err) {
-        if (err) { return done(err); }
-
-        expect(Task.prototype.run).to.be.calledOnce;
-        expect(process.stdout.write).to.be.calledWithMatch("BAR_TASK");
-
-        done();
-      });
-
-    }));
-
-    it("runs with quiet log output", stdioWrap(function (done) {
-      base.sandbox.spy(Task.prototype, "run");
-      base.mockFs({
-        "package.json": JSON.stringify({
-          "scripts": {
-            "bar": "echo BAR_TASK"
+            "bar": "echo BAR_TASK >> stdout.log"
           }
         }, null, 2)
       });
@@ -314,22 +292,22 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.run).to.be.calledOnce;
-        expect(process.stdout.write).to.be.calledWithMatch("BAR_TASK");
         expect(logStubs.info).not.be.called;
         expect(logStubs.warn).not.be.called;
         expect(logStubs.error).not.be.called;
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data).to.contain("BAR_TASK");
+        }, done);
       });
-    }));
+    });
 
-    it("runs with specified log level", stdioWrap(function (done) {
-      base.sandbox.stub(process.stderr, "write");
+    it("runs with specified log level", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         "package.json": JSON.stringify({
           "scripts": {
-            "bar": "BAD_COMMAND"
+            "bar": "BAD_COMMAND 2>> stderr.log"
           }
         }, null, 2)
       });
@@ -342,19 +320,20 @@ describe("bin/builder-core", function () {
           .that.contains("BAD_COMMAND");
 
         expect(Task.prototype.run).to.be.calledOnce;
-        expect(process.stderr.write).to.be.calledWithMatch("BAD_COMMAND");
         expect(logStubs.info).not.be.called;
         expect(logStubs.warn).not.be.called;
         expect(logStubs.error)
           .to.be.calledWithMatch("Command failed").and
           .to.be.calledWithMatch("BAD_COMMAND");
 
-        done();
+        readFile("stderr.log", function (data) {
+          expect(data).to.contain("BAD_COMMAND");
+        }, done);
       });
 
-    }));
+    });
 
-    it("runs an <archetype>/package.json command", stdioWrap(function (done) {
+    it("runs an <archetype>/package.json command", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -363,7 +342,7 @@ describe("bin/builder-core", function () {
           "mock-archetype": {
             "package.json": JSON.stringify({
               "scripts": {
-                "foo": "echo FOO_TASK"
+                "foo": "echo FOO_TASK >> stdout.log"
               }
             }, null, 2)
           }
@@ -376,12 +355,13 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.run).to.be.calledOnce;
-        expect(process.stdout.write).to.be.calledWithMatch("FOO_TASK");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data).to.contain("FOO_TASK");
+        }, done);
       });
 
-    }));
+    });
 
     it("ignores archetype builder:-prefaced tasks", function () {
       base.mockFs({
@@ -414,20 +394,20 @@ describe("bin/builder-core", function () {
       throw new Error("should have already thrown");
     });
 
-    it("overrides a <archetype> command with a <root> one", stdioWrap(function (done) {
+    it("overrides a <archetype> command with a <root> one", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         ".builderrc": "---\narchetypes:\n  - mock-archetype",
         "package.json": JSON.stringify({
           "scripts": {
-            "foo": "echo ROOT_TASK"
+            "foo": "echo ROOT_TASK >> stdout.log"
           }
         }, null, 2),
         "node_modules": {
           "mock-archetype": {
             "package.json": JSON.stringify({
               "scripts": {
-                "foo": "echo ARCH_TASK"
+                "foo": "echo ARCH_TASK >> stdout.log"
               }
             }, null, 2)
           }
@@ -440,12 +420,13 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.run).to.be.calledOnce;
-        expect(process.stdout.write).to.be.calledWithMatch("ROOT_TASK");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data).to.contain("ROOT_TASK");
+        }, done);
       });
 
-    }));
+    });
 
     // TODO: This one is going to be... tough.
     // https://github.com/FormidableLabs/builder/issues/9
@@ -453,14 +434,14 @@ describe("bin/builder-core", function () {
 
     // TODO: Fix flake in --setup tests.
     // https://github.com/FormidableLabs/builder/issues/86
-    it.skip("runs with --setup", stdioWrap(function (done) {
+    it.skip("runs with --setup", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         "package.json": JSON.stringify({
           "scripts": {
             // *real* fs for script references. (`0` runs forever).
-            "setup": "node test/server/fixtures/repeat-script.js 0 SETUP",
-            "bar": "node test/server/fixtures/repeat-script.js 5 BAR_TASK"
+            "setup": "node test/server/fixtures/repeat-script.js 0 SETUP >> stdout.log",
+            "bar": "node test/server/fixtures/repeat-script.js 5 BAR_TASK >> stdout.log"
           }
         }, null, 2)
       });
@@ -471,19 +452,20 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.run).to.have.callCount(2);
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("SETUP").and
-          .to.be.calledWithMatch("BAR_TASK").and
-          .to.be.calledWithMatch("EXIT - BAR_TASK - 0");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data)
+            .to.contain("SETUP").and
+            .to.contain("BAR_TASK").and
+            .to.contain("EXIT - BAR_TASK - 0");
+        }, done);
       });
 
-    }));
+    });
 
     // TODO: Fix flake in --setup tests.
     // https://github.com/FormidableLabs/builder/issues/86
-    it.skip("handles --setup early 0 exit", stdioWrap(function (done) {
+    it.skip("handles --setup early 0 exit", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         "package.json": JSON.stringify({
@@ -508,11 +490,11 @@ describe("bin/builder-core", function () {
         done();
       });
 
-    }));
+    });
 
     // TODO: Fix flake in --setup tests.
     // https://github.com/FormidableLabs/builder/issues/86
-    it.skip("handles --setup early 1 exit", stdioWrap(function (done) {
+    it.skip("handles --setup early 1 exit", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         "package.json": JSON.stringify({
@@ -537,15 +519,14 @@ describe("bin/builder-core", function () {
         done();
       });
 
-    }));
+    });
 
-    it("runs with --tries=2", stdioWrap(function (done) {
-      base.sandbox.stub(process.stderr, "write");
+    it("runs with --tries=2", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         "package.json": JSON.stringify({
           "scripts": {
-            "foo": "BAD_COMMAND"
+            "foo": "BAD_COMMAND 2>> stderr.log"
           }
         }, null, 2)
       });
@@ -558,18 +539,19 @@ describe("bin/builder-core", function () {
           .that.contains("BAD_COMMAND");
 
         expect(Task.prototype.run).to.be.calledOnce;
-        expect(process.stderr.write).to.be.calledWithMatch("BAD_COMMAND");
         expect(logStubs.warn).to.be.calledWithMatch(chalk.red("1") + " tries left");
         expect(logStubs.error)
           .to.be.calledWithMatch("Command failed").and
           .to.be.calledWithMatch("BAD_COMMAND");
 
-        done();
+        readFile("stderr.log", function (data) {
+          expect(data).to.contain("BAD_COMMAND");
+        }, done);
       });
 
-    }));
+    });
 
-    it("runs with base config value", stdioWrap(function (done) {
+    it("runs with base config value", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         "package.json": JSON.stringify({
@@ -577,7 +559,7 @@ describe("bin/builder-core", function () {
             "_test_message": "from base config"
           },
           "scripts": {
-            "echo": "node test/server/fixtures/echo.js"
+            "echo": "node test/server/fixtures/echo.js >> stdout.log"
           }
         }, null, 2)
       });
@@ -588,14 +570,14 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.run).to.have.callCount(1);
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("string - from base config");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data).to.contain("string - from base config");
+        }, done);
       });
-    }));
+    });
 
-    it("runs with archetype config value", stdioWrap(function (done) {
+    it("runs with archetype config value", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -607,7 +589,7 @@ describe("bin/builder-core", function () {
                 "_test_message": "from archetype"
               },
               "scripts": {
-                "echo": "node test/server/fixtures/echo.js"
+                "echo": "node test/server/fixtures/echo.js >> stdout.log"
               }
             }, null, 2)
           }
@@ -620,14 +602,14 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.run).to.have.callCount(1);
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("string - from archetype");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data).to.contain("string - from archetype");
+        }, done);
       });
-    }));
+    });
 
-    it("runs with empty base + non-empty archetype config value", stdioWrap(function (done) {
+    it("runs with empty base + non-empty archetype config value", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -643,7 +625,7 @@ describe("bin/builder-core", function () {
                 "_test_message": "from archetype"
               },
               "scripts": {
-                "echo": "node test/server/fixtures/echo.js"
+                "echo": "node test/server/fixtures/echo.js >> stdout.log"
               }
             }, null, 2)
           }
@@ -656,14 +638,14 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.run).to.have.callCount(1);
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("string - EMPTY");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data).to.contain("string - EMPTY");
+        }, done);
       });
-    }));
+    });
 
-    it("runs with real ENV overriding archetype config value", stdioWrap(function (done) {
+    it("runs with real ENV overriding archetype config value", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -675,7 +657,7 @@ describe("bin/builder-core", function () {
                 "_test_message": "from archetype"
               },
               "scripts": {
-                "echo": "node test/server/fixtures/echo.js"
+                "echo": "node test/server/fixtures/echo.js >> stdout.log"
               }
             }, null, 2)
           }
@@ -692,14 +674,14 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.run).to.have.callCount(1);
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("string - from real env");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data).to.contain("string - from real env");
+        }, done);
       });
-    }));
+    });
 
-    it("runs with real ENV overriding base + archetype config values", stdioWrap(function (done) {
+    it("runs with real ENV overriding base + archetype config values", function (done) {
       base.sandbox.spy(Task.prototype, "run");
       base.mockFs({
         ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -715,7 +697,7 @@ describe("bin/builder-core", function () {
                 "_test_message": "from archetype"
               },
               "scripts": {
-                "echo": "node test/server/fixtures/echo.js"
+                "echo": "node test/server/fixtures/echo.js >> stdout.log"
               }
             }, null, 2)
           }
@@ -732,16 +714,16 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.run).to.have.callCount(1);
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("string - from real env");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data).to.contain("string - from real env");
+        }, done);
       });
-    }));
+    });
 
     describe("expands paths with --expand-archetype", function () {
 
-      it("Skips `../node_modules/<archetype>`", stdioWrap(function (done) {
+      it("Skips `../node_modules/<archetype>`", function (done) {
         base.sandbox.spy(Task.prototype, "run");
         base.mockFs({
           ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -750,7 +732,7 @@ describe("bin/builder-core", function () {
             "mock-archetype": {
               "package.json": JSON.stringify({
                 "scripts": {
-                  "bar": "echo WONT_EXPAND ../node_modules/mock-archetype/A_FILE.txt"
+                  "bar": "echo WONT_EXPAND ../node_modules/mock-archetype/A_FILE.txt >> stdout.log"
                 }
               }, null, 2)
             }
@@ -763,15 +745,16 @@ describe("bin/builder-core", function () {
           if (err) { return done(err); }
 
           expect(Task.prototype.run).to.be.calledOnce;
-          expect(process.stdout.write).to.be.calledWithMatch(
-            "WONT_EXPAND ../node_modules/mock-archetype/A_FILE.txt"
-          );
 
-          done();
+          readFile("stdout.log", function (data) {
+            expect(data).to.contain(
+              "WONT_EXPAND ../node_modules/mock-archetype/A_FILE.txt"
+            );
+          }, done);
         });
-      }));
+      });
 
-      it("Skips `other/node_modules/<archetype>`", stdioWrap(function (done) {
+      it("Skips `other/node_modules/<archetype>`", function (done) {
         base.sandbox.spy(Task.prototype, "run");
         base.mockFs({
           ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -780,7 +763,8 @@ describe("bin/builder-core", function () {
             "mock-archetype": {
               "package.json": JSON.stringify({
                 "scripts": {
-                  "bar": "echo WONT_EXPAND other/node_modules/mock-archetype/A_FILE.txt"
+                  "bar": "echo WONT_EXPAND " +
+                    "other/node_modules/mock-archetype/A_FILE.txt >> stdout.log"
                 }
               }, null, 2)
             }
@@ -793,15 +777,16 @@ describe("bin/builder-core", function () {
           if (err) { return done(err); }
 
           expect(Task.prototype.run).to.be.calledOnce;
-          expect(process.stdout.write).to.be.calledWithMatch(
-            "WONT_EXPAND other/node_modules/mock-archetype/A_FILE.txt"
-          );
 
-          done();
+          readFile("stdout.log", function (data) {
+            expect(data).to.contain(
+              "WONT_EXPAND other/node_modules/mock-archetype/A_FILE.txt"
+            );
+          }, done);
         });
-      }));
+      });
 
-      it("Replaces `node_modules/<archetype>`", stdioWrap(function (done) {
+      it("Replaces `node_modules/<archetype>`", function (done) {
         base.sandbox.spy(Task.prototype, "run");
         base.mockFs({
           ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -810,7 +795,7 @@ describe("bin/builder-core", function () {
             "mock-archetype": {
               "package.json": JSON.stringify({
                 "scripts": {
-                  "bar": "echo EXPANDED node_modules/mock-archetype/A_FILE.txt"
+                  "bar": "echo EXPANDED node_modules/mock-archetype/A_FILE.txt >> stdout.log"
                 }
               }, null, 2)
             }
@@ -823,15 +808,16 @@ describe("bin/builder-core", function () {
           if (err) { return done(err); }
 
           expect(Task.prototype.run).to.be.calledOnce;
-          expect(process.stdout.write).to.be.calledWithMatch(
-            "EXPANDED " + path.join(process.cwd(), "node_modules/mock-archetype/A_FILE.txt")
-          );
 
-          done();
+          readFile("stdout.log", function (data) {
+            expect(data).to.contain(
+              "EXPANDED " + path.join(process.cwd(), "node_modules/mock-archetype/A_FILE.txt")
+            );
+          }, done);
         });
-      }));
+      });
 
-      it("Replaces `\"node_modules/<archetype>`", stdioWrap(function (done) {
+      it("Replaces `\"node_modules/<archetype>`", function (done) {
         base.sandbox.spy(Task.prototype, "run");
         base.mockFs({
           ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -842,7 +828,8 @@ describe("bin/builder-core", function () {
                 "scripts": {
                   // Note: Need double escaping here for command line echo.
                   // Note: Not sure if the quote escaping is the same on windows.
-                  "bar": "echo EXPANDED \\\"node_modules/mock-archetype/A_FILE.txt\\\""
+                  "bar": "echo EXPANDED " +
+                    "\\\"node_modules/mock-archetype/A_FILE.txt\\\" >> stdout.log"
                 }
               }, null, 2)
             }
@@ -855,15 +842,16 @@ describe("bin/builder-core", function () {
           if (err) { return done(err); }
 
           expect(Task.prototype.run).to.be.calledOnce;
-          expect(process.stdout.write).to.be.calledWithMatch(
-            "EXPANDED \"" + path.join(process.cwd(), "node_modules/mock-archetype/A_FILE.txt\"")
-          );
 
-          done();
+          readFile("stdout.log", function (data) {
+            expect(data).to.contain(
+              "EXPANDED \"" + path.join(process.cwd(), "node_modules/mock-archetype/A_FILE.txt\"")
+            );
+          }, done);
         });
-      }));
+      });
 
-      it("Propagates flag to sub-task", stdioWrap(function (done) {
+      it("Propagates flag to sub-task", function (done) {
         base.sandbox.spy(Task.prototype, "run");
         base.mockFs({
           ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -872,7 +860,7 @@ describe("bin/builder-core", function () {
             "mock-archetype": {
               "package.json": JSON.stringify({
                 "scripts": {
-                  "bar": "echo EXPANDED node_modules/mock-archetype/A_FILE.txt"
+                  "bar": "echo EXPANDED node_modules/mock-archetype/A_FILE.txt >> stdout.log"
                 }
               }, null, 2)
             }
@@ -887,21 +875,22 @@ describe("bin/builder-core", function () {
           if (err) { return done(err); }
 
           expect(Task.prototype.run).to.be.calledOnce;
-          expect(process.stdout.write).to.be.calledWithMatch(
-            "EXPANDED " + path.join(process.cwd(), "node_modules/mock-archetype/A_FILE.txt")
-          );
 
-          done();
+          readFile("stdout.log", function (data) {
+            expect(data).to.contain(
+              "EXPANDED " + path.join(process.cwd(), "node_modules/mock-archetype/A_FILE.txt")
+            );
+          }, done);
         });
-      }));
+      });
 
-      it("Skips replacing root project tasks", stdioWrap(function (done) {
+      it("Skips replacing root project tasks", function (done) {
         base.sandbox.spy(Task.prototype, "run");
         base.mockFs({
           ".builderrc": "---\narchetypes:\n  - mock-archetype",
           "package.json": JSON.stringify({
             "scripts": {
-              "bar": "echo WONT_EXPAND node_modules/mock-archetype/A_FILE.txt"
+              "bar": "echo WONT_EXPAND node_modules/mock-archetype/A_FILE.txt >> stdout.log"
             }
           }, null, 2),
           "node_modules": {
@@ -917,13 +906,14 @@ describe("bin/builder-core", function () {
           if (err) { return done(err); }
 
           expect(Task.prototype.run).to.be.calledOnce;
-          expect(process.stdout.write).to.be.calledWithMatch(
-            "WONT_EXPAND node_modules/mock-archetype/A_FILE.txt"
-          );
 
-          done();
+          readFile("stdout.log", function (data) {
+            expect(data).to.contain(
+              "WONT_EXPAND node_modules/mock-archetype/A_FILE.txt"
+            );
+          }, done);
         });
-      }));
+      });
 
     });
 
@@ -931,14 +921,14 @@ describe("bin/builder-core", function () {
 
   describe("builder concurrent", function () {
 
-    it("runs <root>/package.json concurrent commands", stdioWrap(function (done) {
+    it("runs <root>/package.json concurrent commands", function (done) {
       base.sandbox.spy(Task.prototype, "concurrent");
       base.mockFs({
         "package.json": JSON.stringify({
           "scripts": {
-            "one": "echo ONE_TASK",
-            "two": "echo TWO_TASK",
-            "three": "echo THREE_TASK"
+            "one": "echo ONE_TASK >> stdout.log",
+            "two": "echo TWO_TASK >> stdout.log",
+            "three": "echo THREE_TASK >> stdout.log"
           }
         }, null, 2)
       });
@@ -949,33 +939,34 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.concurrent).to.be.calledOnce;
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("ONE_TASK").and
-          .to.be.calledWithMatch("TWO_TASK").and
-          .to.be.calledWithMatch("THREE_TASK");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data)
+            .to.contain("ONE_TASK").and
+            .to.contain("TWO_TASK").and
+            .to.contain("THREE_TASK");
+        }, done);
       });
 
-    }));
+    });
 
-    it("runs <archetype>/package.json concurrent commands", stdioWrap(function (done) {
+    it("runs <archetype>/package.json concurrent commands", function (done) {
       base.sandbox.spy(Task.prototype, "concurrent");
       base.mockFs({
         ".builderrc": "---\narchetypes:\n  - mock-archetype",
         "package.json": JSON.stringify({
           "scripts": {
-            "two": "echo TWO_ROOT_TASK",
-            "three": "echo THREE_ROOT_TASK"
+            "two": "echo TWO_ROOT_TASK >> stdout.log",
+            "three": "echo THREE_ROOT_TASK >> stdout.log"
           }
         }, null, 2),
         "node_modules": {
           "mock-archetype": {
             "package.json": JSON.stringify({
               "scripts": {
-                "one": "echo ONE_TASK",
-                "two": "echo TWO_TASK",
-                "three": "echo THREE_TASK"
+                "one": "echo ONE_TASK >> stdout.log",
+                "two": "echo TWO_TASK >> stdout.log",
+                "three": "echo THREE_TASK >> stdout.log"
               }
             }, null, 2)
           }
@@ -988,15 +979,16 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.concurrent).to.be.calledOnce;
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("ONE_TASK").and
-          .to.be.calledWithMatch("TWO_ROOT_TASK").and
-          .to.be.calledWithMatch("THREE_ROOT_TASK");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data)
+            .to.contain("ONE_TASK").and
+            .to.contain("TWO_ROOT_TASK").and
+            .to.contain("THREE_ROOT_TASK");
+        }, done);
       });
 
-    }));
+    });
 
     // TODO: Finish outlined tests.
     // https://github.com/FormidableLabs/builder/issues/9
@@ -1004,7 +996,7 @@ describe("bin/builder-core", function () {
     it("runs with --setup");
     it("runs with --queue=1, --bail=false");
 
-    it("runs with base overriding archetype config value", stdioWrap(function (done) {
+    it("runs with base overriding archetype config value", function (done) {
       base.sandbox.spy(Task.prototype, "concurrent");
       base.mockFs({
         ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -1013,7 +1005,7 @@ describe("bin/builder-core", function () {
             "_test_message": "from base"
           },
           "scripts": {
-            "echo2": "node test/server/fixtures/echo.js TWO"
+            "echo2": "node test/server/fixtures/echo.js TWO >> stdout.log"
           }
         }, null, 2),
         "node_modules": {
@@ -1023,7 +1015,7 @@ describe("bin/builder-core", function () {
                 "_test_message": "from archetype"
               },
               "scripts": {
-                "echo1": "node test/server/fixtures/echo.js ONE"
+                "echo1": "node test/server/fixtures/echo.js ONE >> stdout.log"
               }
             }, null, 2)
           }
@@ -1036,24 +1028,26 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.concurrent).to.have.callCount(1);
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("string - from base - ONE").and
-          .to.be.calledWithMatch("string - from base - TWO");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data)
+            .to.contain("string - from base - ONE").and
+            .to.contain("string - from base - TWO");
+        }, done);
       });
-    }));
+    });
 
   });
 
   describe("builder envs", function () {
 
-    it("runs <root>/package.json multiple env commands", stdioWrap(function (done) {
+    it("runs <root>/package.json multiple env commands", function (done) {
       base.sandbox.spy(Task.prototype, "envs");
       base.mockFs({
         "package.json": JSON.stringify({
           "scripts": {
-            "echo": "echo ROOT " + (/^win/.test(process.platform) ? "%MY_VAR%" : "$MY_VAR")
+            "echo": "echo ROOT " + (/^win/.test(process.platform) ? "%MY_VAR%" : "$MY_VAR") +
+              " >> stdout.log"
           }
         }, null, 2)
       });
@@ -1068,22 +1062,24 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.envs).to.be.calledOnce;
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("ROOT hi").and
-          .to.be.calledWithMatch("ROOT ho").and
-          .to.be.calledWithMatch("ROOT yo");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data)
+            .to.contain("ROOT hi").and
+            .to.contain("ROOT ho").and
+            .to.contain("ROOT yo");
+        }, done);
       });
 
-    }));
+    });
 
-    it("runs multiple env commands with --buffer, --envs-path", stdioWrap(function (done) {
+    it("runs multiple env commands with --buffer, --envs-path", function (done) {
       base.sandbox.spy(Task.prototype, "envs");
       base.mockFs({
         "package.json": JSON.stringify({
           "scripts": {
-            "echo": "echo ROOT " + (/^win/.test(process.platform) ? "%MY_VAR%" : "$MY_VAR")
+            "echo": "echo ROOT " + (/^win/.test(process.platform) ? "%MY_VAR%" : "$MY_VAR") +
+              " >> stdout.log"
           }
         }, null, 2),
         "envs.json": JSON.stringify([
@@ -1099,17 +1095,18 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.envs).to.be.calledOnce;
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("ROOT hi").and
-          .to.be.calledWithMatch("ROOT ho").and
-          .to.be.calledWithMatch("ROOT yo");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data)
+            .to.contain("ROOT hi").and
+            .to.contain("ROOT ho").and
+            .to.contain("ROOT yo");
+        }, done);
       });
 
-    }));
+    });
 
-    it("runs <archetype>/package.json multiple env commands", stdioWrap(function (done) {
+    it("runs <archetype>/package.json multiple env commands", function (done) {
       base.sandbox.spy(Task.prototype, "envs");
       base.mockFs({
         ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -1118,7 +1115,8 @@ describe("bin/builder-core", function () {
           "mock-archetype": {
             "package.json": JSON.stringify({
               "scripts": {
-                "echo": "echo ARCH " + (/^win/.test(process.platform) ? "%MY_VAR%" : "$MY_VAR")
+                "echo": "echo ARCH " + (/^win/.test(process.platform) ? "%MY_VAR%" : "$MY_VAR") +
+                  " >> stdout.log"
               }
             }, null, 2)
           }
@@ -1135,30 +1133,33 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.envs).to.be.calledOnce;
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("ARCH hi").and
-          .to.be.calledWithMatch("ARCH ho").and
-          .to.be.calledWithMatch("ARCH yo");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data)
+            .to.contain("ARCH hi").and
+            .to.contain("ARCH ho").and
+            .to.contain("ARCH yo");
+        }, done);
       });
 
-    }));
+    });
 
-    it("overrides <archetype>/package.json multiple env commands", stdioWrap(function (done) {
+    it("overrides <archetype>/package.json multiple env commands", function (done) {
       base.sandbox.spy(Task.prototype, "envs");
       base.mockFs({
         ".builderrc": "---\narchetypes:\n  - mock-archetype",
         "package.json": JSON.stringify({
           "scripts": {
-            "echo": "echo ROOT " + (/^win/.test(process.platform) ? "%MY_VAR%" : "$MY_VAR")
+            "echo": "echo ROOT " + (/^win/.test(process.platform) ? "%MY_VAR%" : "$MY_VAR") +
+              " >> stdout.log"
           }
         }, null, 2),
         "node_modules": {
           "mock-archetype": {
             "package.json": JSON.stringify({
               "scripts": {
-                "echo": "echo ARCH " + (/^win/.test(process.platform) ? "%MY_VAR%" : "$MY_VAR")
+                "echo": "echo ARCH " + (/^win/.test(process.platform) ? "%MY_VAR%" : "$MY_VAR") +
+                  " >> stdout.log"
               }
             }, null, 2)
           }
@@ -1175,17 +1176,18 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.envs).to.be.calledOnce;
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("ROOT hi").and
-          .to.be.calledWithMatch("ROOT ho").and
-          .to.be.calledWithMatch("ROOT yo");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data)
+            .to.contain("ROOT hi").and
+            .to.contain("ROOT ho").and
+            .to.contain("ROOT yo");
+        }, done);
       });
 
-    }));
+    });
 
-    it("errors on empty JSON array", stdioWrap(function (done) {
+    it("errors on empty JSON array", function (done) {
       base.sandbox.spy(Task.prototype, "envs");
       base.mockFs({
         "package.json": JSON.stringify({
@@ -1205,9 +1207,9 @@ describe("bin/builder-core", function () {
         done();
       });
 
-    }));
+    });
 
-    it("errors on empty JSON non-array", stdioWrap(function (done) {
+    it("errors on empty JSON non-array", function (done) {
       base.sandbox.spy(Task.prototype, "envs");
       base.mockFs({
         "package.json": JSON.stringify({
@@ -1227,9 +1229,9 @@ describe("bin/builder-core", function () {
         done();
       });
 
-    }));
+    });
 
-    it("errors on malformed JSON array", stdioWrap(function (done) {
+    it("errors on malformed JSON array", function (done) {
       base.sandbox.spy(Task.prototype, "envs");
       base.mockFs({
         "package.json": JSON.stringify({
@@ -1251,9 +1253,9 @@ describe("bin/builder-core", function () {
 
         done();
       });
-    }));
+    });
 
-    it("errors on nonexistent JSON file", stdioWrap(function (done) {
+    it("errors on nonexistent JSON file", function (done) {
       base.sandbox.spy(Task.prototype, "envs");
       base.mockFs({
         "package.json": JSON.stringify({
@@ -1275,7 +1277,7 @@ describe("bin/builder-core", function () {
 
         done();
       });
-    }));
+    });
 
     // TODO: Finish outlined tests.
     // https://github.com/FormidableLabs/builder/issues/9
@@ -1283,7 +1285,7 @@ describe("bin/builder-core", function () {
     it("runs with --tries=2");
     it("runs with --queue=1, --bail=false");
 
-    it("runs with envs overriding base config value", stdioWrap(function (done) {
+    it("runs with envs overriding base config value", function (done) {
       base.sandbox.spy(Task.prototype, "envs");
       base.mockFs({
         "package.json": JSON.stringify({
@@ -1291,7 +1293,7 @@ describe("bin/builder-core", function () {
             "_test_message": "from base"
           },
           "scripts": {
-            "echo": "node test/server/fixtures/echo.js"
+            "echo": "node test/server/fixtures/echo.js >> stdout.log"
           }
         }, null, 2)
       });
@@ -1306,16 +1308,17 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.envs).to.have.callCount(1);
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("string - from base").and
-          .to.be.calledWithMatch("string - from array1").and
-          .to.be.calledWithMatch("string - from array2");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data)
+            .to.contain("string - from base").and
+            .to.contain("string - from array1").and
+            .to.contain("string - from array2");
+        }, done);
       });
-    }));
+    });
 
-    it("runs with envs real ENV overriding base + arch config value", stdioWrap(function (done) {
+    it("runs with envs real ENV overriding base + arch config value", function (done) {
       base.sandbox.spy(Task.prototype, "envs");
       base.mockFs({
         ".builderrc": "---\narchetypes:\n  - mock-archetype",
@@ -1331,7 +1334,7 @@ describe("bin/builder-core", function () {
                 "_test_message": "from archetype"
               },
               "scripts": {
-                "echo": "node test/server/fixtures/echo.js"
+                "echo": "node test/server/fixtures/echo.js >> stdout.log"
               }
             }, null, 2)
           }
@@ -1352,14 +1355,15 @@ describe("bin/builder-core", function () {
         if (err) { return done(err); }
 
         expect(Task.prototype.envs).to.have.callCount(1);
-        expect(process.stdout.write)
-          .to.be.calledWithMatch("string - from real env").and
-          .to.be.calledWithMatch("string - from array1").and
-          .to.be.calledWithMatch("string - from array2");
 
-        done();
+        readFile("stdout.log", function (data) {
+          expect(data)
+            .to.contain("string - from real env").and
+            .to.contain("string - from array1").and
+            .to.contain("string - from array2");
+        }, done);
       });
-    }));
+    });
 
   });